Add commands to search more easily from within vim
Now the `s` key is remapped to start a motion to capture and search
text. E.g. `stf` will search (using fzf.vim's `:Ag`) from the from the cursor to the
next occurrence of the `f` character.

This also adds a command (mapped to `<leader>e`) to search (using
fzf.vim's `:Ag`) but ignoring the matches occurring in the file name.